Establish bare bone tests target for HTMLParser
There are likely more files that could go into this target, but in the
context of getting `HTMLParser` accessing by WooCommerce on watchOS
(without UIKit) it's best to move forward without clearly separating the
tests.

The tests in Aztec already exercise the logic in the package anyway, so
we remain covered.